Organize new file paths and complete router implementation
